#770 - Moved stylesheets to appropriate places and refactored imports.
