+++ src/frontend/layouts/nav-top-bar/nav-top-bar.tsx
@@ -12,7 +12,7 @@ import { useTheme } from '../../../services/theme.hooks';
 import themes from '../../../shared/themes';
 import NavUserMenu from './nav-user-menu/nav-user-menu';
 import NavNotificationsMenu from './nav-notifications-menu/nav-notifications-menu';
-import styles from './nav-top-bar.module.css';
+import styles from '../../stylesheets/layouts/nav-top-bar/nav-top-bar.module.css';
 
 const NavTopBar: React.FC = () => {
   const auth = useAuth();
