+++ src/frontend/layouts/page-title/page-breadcrumbs/page-breadcrumbs.tsx
@@ -7,7 +7,7 @@ import { Link } from 'react-router-dom';
 import { Breadcrumb } from 'react-bootstrap';
 import { LinkItem } from '../../../../shared/types';
 import { routes } from '../../../../shared/routes';
-import styles from './page-breadcrumbs.module.css';
+import styles from '../../../stylesheets/layouts/page-breadcrumbs.module.css';
 
 interface PageTitleProps {
   currentPageTitle: string;
