+++ src/frontend/layouts/sidebar/sidebar.tsx
@@ -13,7 +13,7 @@ import {
 import { routes } from '../../../shared/routes';
 import { LinkItem } from '../../../shared/types';
 import NavPageLinks from './nav-page-links/nav-page-links';
-import styles from './sidebar.module.css';
+import styles from '../../stylesheets/layouts/sidebar/sidebar.module.css';
 
 const Sidebar: React.FC = () => {
   const linkItems: LinkItem[] = [
