+++ src/frontend/pages/ErrorPage/error-page.tsx
@@ -5,7 +5,7 @@
 
 import { Container } from 'react-bootstrap';
 import PageBlock from '../../layouts/page-block/page-block';
-import styles from './error-page.module.css';
+import styles from '../../stylesheets/pages/error-page.module.css';
 
 interface ErrorPageProps {
   message?: string;
