+++ src/frontend/pages/ProjectDetailPage/project-edit-container/edit-mode-options/edit-mode-options.tsx
@@ -4,7 +4,7 @@
  */
 
 import { Button } from 'react-bootstrap';
-import styles from './edit-mode-options.module.css';
+import styles from '../../../../stylesheets/pages/ProjectDetailPage/edit-mode-options.module.css';
 
 interface EditModeOptionsProps {
   exitEditMode: () => void;
