+++ src/frontend/stylesheets/pages/ProjectDetailPage/edit-mode-options.module.css
