+++ src/frontend/stylesheets/pages/ProjectDetailPage/work-package-summary.module.scss
