+++ src/frontend/stylesheets/pages/projects-table-filter.module.scss
